+++ examples/Example1.java
@@ -15,25 +15,23 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the MOEA Framework.  If not, see <http://www.gnu.org/licenses/>.
  */
-import org.moeaframework.Executor;
-import org.moeaframework.core.NondominatedPopulation;
+import org.moeaframework.algorithm.NSGAII;
+import org.moeaframework.core.Problem;
+import org.moeaframework.problem.DTLZ.DTLZ2;
 
 /**
- * Demonstrates using an Executor to solve the UF1 test problem with NSGA-II,
- * one of the most widely-used multiobjective evolutionary algorithms.
+ * Simplest way to solve a multi-objective optimization problem.  Here, we solve
+ * the bi-objective DTLZ2 test problem using NSGA-II and display the Pareto front.
  */
 public class Example1 {
 
 	public static void main(String[] args) {
-		//configure and run this experiment
-		NondominatedPopulation result = new Executor()
-				.withProblem("UF1")
-				.withAlgorithm("NSGAII")
-				.withMaxEvaluations(10000)
-				.run();
+		Problem problem = new DTLZ2(2);
 		
-		//display the results
-		result.display();
+		NSGAII algorithm = new NSGAII(problem);
+		algorithm.run(10000);
+		
+		algorithm.getResult().display();
 	}
 
 }