+++ examples/org/moeaframework/examples/configuring/GetAndSetConfiguration.java
@@ -21,7 +21,6 @@ import java.io.IOException;
 
 import org.moeaframework.algorithm.NSGAII;
 import org.moeaframework.core.Problem;
-import org.moeaframework.core.Solution;
 import org.moeaframework.problem.DTLZ.DTLZ2;
 import org.moeaframework.util.TypedProperties;
 
@@ -53,13 +52,8 @@ public class GetAndSetConfiguration {
 		algorithm.run(10000);
 		
 		System.out.println();
-		System.out.println("Objective1  Objective2");
 		
-		for (Solution solution : algorithm.getResult()) {
-			System.out.format("%.4f      %.4f%n",
-					solution.getObjective(0),
-					solution.getObjective(1));
-		}
+		algorithm.getResult().display();
 	}
 
 }